requirements.txt
boto3==1.40.5
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
matplotlib==3.6.3
numpy==1.26.4
openpyxl==3.1.5
pandas==2.3.1
pytest==8.4.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.7.1
python-multipart==0.0.20
